# frontend/src/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "smd-web-hmi 后端",
    "version": "0.3.0-rc.5"
  },
  "paths": {
...
    "/api/reports": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "List Reports",
        "description": "报告列表。权限：Observer+。",
        "operationId": "list_reports_api_reports_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/generate": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Generate Report",
        "description": "提交报告生成任务并立即返回。权限：Operator+。",
        "operationId": "generate_report_api_reports_generate_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateReportRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/tasks/{task_id}": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Report Task Status",
        "description": "查询报告生成任务状态。",
        "operationId": "report_task_status_api_reports_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 32,
              "maxLength": 32,
              "pattern": "^[0-9a-f]{32}$",
              "title": "Task Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/{report_id}/download": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Download Report",
        "description": "下载报告文件。权限：Observer+。",
        "operationId": "download_report_api_reports__report_id__download_get",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Report Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/run-recoveries": {
      "get": {
        "tags": [
          "run recovery"
        ],
        "summary": "List Recoveries",
        "operationId": "list_recoveries_api_run_recoveries_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryResponse_list_RecoverySummary__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/run-recoveries/{recovery_id}": {
      "get": {
        "tags": [
          "run recovery"
        ],
        "summary": "Recovery Detail",
        "operationId": "recovery_detail_api_run_recoveries__recovery_id__get",
        "parameters": [
          {
            "name": "recovery_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Recovery Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryResponse_RecoveryDetail_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/run-recoveries/{recovery_id}/binding": {
      "post": {
        "tags": [
          "run recovery"
        ],
        "summary": "Bind Recovery",
        "operationId": "bind_recovery_api_run_recoveries__recovery_id__binding_post",
        "parameters": [
          {
            "name": "recovery_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Recovery Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoveryBindingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryResponse_RecoverySummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/run-recoveries/{recovery_id}/replays": {
      "post": {
        "tags": [
          "run recovery"
        ],
        "summary": "Replay Recovery",
        "operationId": "replay_recovery_api_run_recoveries__recovery_id__replays_post",
        "parameters": [
          {
            "name": "recovery_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Recovery Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoveryReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryResponse_RecoverySummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenerateReportRequest": {
        "properties": {
          "test_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Test Id"
          },
          "format": {
            "type": "string",
            "pattern": "^(html|pdf|xlsx)$",
            "title": "Format",
            "default": "html"
          },
          "options": {
            "type": "object",
            "maxProperties": 20,
            "title": "Options"
          }
        },
        "type": "object",
        "required": [
          "test_id"
        ],
        "title": "GenerateReportRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RecoveryAudit": {
        "properties": {
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "evidence": {
            "type": "object",
            "title": "Evidence"
          }
        },
        "type": "object",
        "required": [
          "actor",
          "role",
          "reason",
          "action",
          "created_at",
          "evidence"
        ],
        "title": "RecoveryAudit"
      },
      "RecoveryBindingRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 2,
            "title": "Reason"
          },
          "idempotency_key": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Idempotency Key"
          },
          "expected_review_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Review Revision"
          },
          "target_test_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Test Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reason",
          "idempotency_key",
          "expected_review_revision"
        ],
        "title": "RecoveryBindingRequest"
      },
      "RecoveryDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "readOnly": true
          },
          "device_id": {
            "type": "string",
            "title": "Device Id",
            "readOnly": true
          },
          "run_id": {
            "type": "string",
            "title": "Run Id",
            "readOnly": true
          },
          "first_seen_at": {
            "type": "string",
            "title": "First Seen At"
          },
          "last_seen_at": {
            "type": "string",
            "title": "Last Seen At"
          },
          "review_revision": {
            "type": "integer",
            "title": "Review Revision"
          },
          "review_state": {
            "type": "string",
            "enum": [
              "unreviewed",
              "bound",
              "conflict"
            ],
            "title": "Review State"
          },
          "test_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Test Id"
          },
          "replay_status": {
            "type": "string",
            "enum": [
              "not_bound",
              "pending",
              "running",
              "complete",
              "failed"
            ],
            "title": "Replay Status"
          },
          "replay_through_id": {
            "type": "integer",
            "title": "Replay Through Id"
          },
          "replay_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replay Error"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "evidence": {
            "type": "object",
            "title": "Evidence"
          },
          "start_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "discovered_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discovered At"
          },
          "source_count": {
            "type": "integer",
            "title": "Source Count"
          },
          "log_gaps": {
            "items": {
              "$ref": "#/components/schemas/RecoveryLogGap"
            },
            "type": "array",
            "title": "Log Gaps"
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/RecoveryAudit"
            },
            "type": "array",
            "title": "Reviews"
          }
        },
        "type": "object",
        "required": [
          "id",
          "device_id",
          "run_id",
          "first_seen_at",
          "last_seen_at",
          "review_revision",
          "review_state",
          "test_id",
          "replay_status",
          "replay_through_id",
          "replay_error",
          "state",
          "evidence",
          "start_time",
          "discovered_at",
          "source_count",
          "log_gaps",
          "reviews"
        ],
        "title": "RecoveryDetail"
      },
      "RecoveryLogGap": {
        "properties": {
          "log_id": {
            "type": "string",
            "title": "Log Id"
          },
          "first_record_seq": {
            "type": "string",
            "title": "First Record Seq"
          },
          "last_record_seq": {
            "type": "string",
            "title": "Last Record Seq"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "log_id",
          "first_record_seq",
          "last_record_seq",
          "reason"
        ],
        "title": "RecoveryLogGap"
      },
      "RecoveryResponse_RecoveryDetail_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RecoveryDetail"
          },
          "ts": {
            "type": "string",
            "title": "Ts"
          }
        },
        "type": "object",
        "required": [
          "data",
          "ts"
        ],
        "title": "RecoveryResponse[RecoveryDetail]"
      },
      "RecoveryResponse_RecoverySummary_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RecoverySummary"
          },
          "ts": {
            "type": "string",
            "title": "Ts"
          }
        },
        "type": "object",
        "required": [
          "data",
          "ts"
        ],
        "title": "RecoveryResponse[RecoverySummary]"
      },
      "RecoveryResponse_list_RecoverySummary__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/RecoverySummary"
            },
            "type": "array",
            "title": "Data"
          },
          "ts": {
            "type": "string",
            "title": "Ts"
          }
        },
        "type": "object",
        "required": [
          "data",
          "ts"
        ],
        "title": "RecoveryResponse[list[RecoverySummary]]"
      },
      "RecoveryReviewRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 2,
            "title": "Reason"
          },
          "idempotency_key": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Idempotency Key"
          },
          "expected_review_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Review Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reason",
          "idempotency_key",
          "expected_review_revision"
        ],
        "title": "RecoveryReviewRequest"
      },
      "RecoverySummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "readOnly": true
          },
          "device_id": {
            "type": "string",
            "title": "Device Id",
            "readOnly": true
          },
          "run_id": {
            "type": "string",
            "title": "Run Id",
            "readOnly": true
          },
          "first_seen_at": {
            "type": "string",
            "title": "First Seen At"
          },
          "last_seen_at": {
            "type": "string",
            "title": "Last Seen At"
          },
          "review_revision": {
            "type": "integer",
            "title": "Review Revision"
          },
          "review_state": {
            "type": "string",
            "enum": [
              "unreviewed",
              "bound",
              "conflict"
            ],
            "title": "Review State"
          },
          "test_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Test Id"
          },
          "replay_status": {
            "type": "string",
            "enum": [
              "not_bound",
              "pending",
              "running",
              "complete",
              "failed"
            ],
            "title": "Replay Status"
          },
          "replay_through_id": {
            "type": "integer",
            "title": "Replay Through Id"
          },
          "replay_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replay Error"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "evidence": {
            "type": "object",
            "title": "Evidence"
          }
        },
        "type": "object",
        "required": [
          "id",
          "device_id",
          "run_id",
          "first_seen_at",
          "last_seen_at",
          "review_revision",
          "review_state",
          "test_id",
          "replay_status",
          "replay_through_id",
          "replay_error",
          "state",
          "evidence"
        ],
        "title": "RecoverySummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}